statistic/requirements.txt
fastapi==0.101.0
uvicorn==0.23.2
grpcio==1.57.0
grpcio-tools==1.57.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.9
